refactor(daemon): distribute configs
