+++ crates/daemon/src/hook/mcp/config.rs
