+++ crates/daemon/src/hook/task/mod.rs
@@ -8,8 +8,10 @@ use tokio::sync::{oneshot, watch};
 use tokio::task::AbortHandle;
 use tokio::time::Instant;
 
+pub use config::TasksConfig;
 pub use registry::TaskRegistry;
 
+pub mod config;
 pub(crate) mod params;
 mod registry;
 pub(crate) mod tool;
