+++ crates/model/src/config.rs
@@ -7,6 +7,7 @@
 use anyhow::{Result, bail};
 use compact_str::CompactString;
 use serde::{Deserialize, Serialize};
+use std::collections::BTreeMap;
 
 /// API protocol standard for remote providers.
 ///
@@ -92,6 +93,20 @@ pub enum Loader {
     Gguf,
     /// `VisionModelBuilder` — vision-language models.
     Vision,
+}
+
+/// Model provider configuration for the daemon.
+///
+/// Remote providers are configured in `providers`. Local models come from the
+/// built-in registry. The active model name lives in `[walrus].model`.
+#[derive(Debug, Clone, Serialize, Deserialize, Default)]
+pub struct ModelConfig {
+    /// Optional embedding model
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub embedding: Option<CompactString>,
+    /// Remote providers (local models come from the built-in registry)
+    #[serde(default)]
+    pub providers: BTreeMap<CompactString, ProviderConfig>,
 }
 
 // MemoryThreshold is generated by build.rs from registry TOML files.
