+++ crates/model/src/lib.rs
@@ -17,7 +17,7 @@ pub mod remote;
 #[path = "../local/mod.rs"]
 pub mod local;
 
-pub use config::{ApiStandard, ProviderConfig};
+pub use config::{ApiStandard, ModelConfig, ProviderConfig};
 pub use manager::ProviderManager;
 pub use provider::{Provider, build_provider};
 pub use reqwest::Client;
